+++ src/cargo/core/compiler/compilation.rs
@@ -22,6 +22,7 @@ pub struct Doctest {
 /// A structure returning the result of a compilation.
 pub struct Compilation<'cfg> {
     /// An array of all tests created during this compilation.
+    /// `(package, target, path_to_test_exe)`
     pub tests: Vec<(Package, Target, PathBuf)>,
 
     /// An array of all binaries created.