+++ tests/testsuite/rustc_info_cache.rs
@@ -4,6 +4,11 @@ use std::env;
 
 #[cargo_test]
 fn rustc_info_cache() {
+    // FIXME: when pipelining rides to stable, enable this test on all channels.
+    if !crate::support::is_nightly() {
+        return;
+    }
+
     let p = project()
         .file("src/main.rs", r#"fn main() { println!("hello"); }"#)
         .build();